// sdk/src/types/claimdata.d.ts
/* eslint-disable */
/**
 * This file was automatically generated by json-schema-to-typescript.
 * DO NOT MODIFY IT BY HAND. Instead, modify the source JSONSchema file,
 * and run json-schema-to-typescript to regenerate this file.
 */

/**
 * Properties of an impact claim
 */
export interface HypercertClaimdata {
  /**
   * Scopes of impact
   */
  impact_scope: {
    name?: string;
    value?: string[];
    excludes?: string[];
    display_value?: string;
    [k: string]: unknown;
  };
  /**
   * Scopes of work
   */
  work_scope: {
    name?: string;
    value?: string[];
    excludes?: string[];
    display_value?: string;
    [k: string]: unknown;
  };
  /**
   * Work time period. The value is UNIX time in seconds from epoch.
   */
  work_timeframe: {
    name?: string;
    value?: number[];
    display_value?: string;
    [k: string]: unknown;
  };
  /**
   * Impact time period. The value is UNIX time in seconds from epoch.
   */
  impact_timeframe: {
    name?: string;
    value?: number[];
    display_value?: string;
    [k: string]: unknown;
  };
  /**
   * Contributors
   */
  contributors: {
    name?: string;
    value?: string[];
    display_value?: string;
    [k: string]: unknown;
  };
  /**
   * Rights
   */
  rights?: {
    name?: string;
    value?: string[];
    excludes?: string[];
    display_value?: string;
    [k: string]: unknown;
  };
  [k: string]: unknown;
}