// sdk/src/types/metadata.d.ts
/* eslint-disable */
/**
 * This file was automatically generated by json-schema-to-typescript.
 * DO NOT MODIFY IT BY HAND. Instead, modify the source JSONSchema file,
 * and run json-schema-to-typescript to regenerate this file.
 */

/**
 * Claim data for hypercert. ERC1155 Metadata compliant
 */
export interface HypercertMetadata {
  /**
   * Identifies the asset to which this token represents
   */
  name: string;
  /**
   * Describes the asset to which this token represents
   */
  description: string;
  /**
   * An url pointing to the external website of the project
   */
  external_url?: string;
  /**
   * A URI pointing to a resource with mime type image/* representing the asset to which this token represents. Consider making any images at a width between 320 and 1080 pixels and aspect ratio between 1.91:1 and 4:5 inclusive.
   */
  image: string;
  /**
   * The version of Hypercert schema used to describe this hypercert
   */
  version?: string;
  /**
   * Describes the asset to which this token represents
   */
  ref?: string;
  /**
   * A CID pointer to the merke tree proof json on ipfs
   */
  allowList?: string;
  properties?: {
    trait_type?: string;
    value?: string;
    [k: string]: unknown;
  }[];
  hypercert?: HypercertClaimdata;
}
/**
 * Properties of an impact claim
 */
export interface HypercertClaimdata {
  /**
   * Scopes of impact
   */
  impact_scope: {
    name?: string;
    value?: string[];
    excludes?: string[];
    display_value?: string;
    [k: string]: unknown;
  };
  /**
   * Scopes of work
   */
  work_scope: {
    name?: string;
    value?: string[];
    excludes?: string[];
    display_value?: string;
    [k: string]: unknown;
  };
  /**
   * Work time period. The value is UNIX time in seconds from epoch.
   */
  work_timeframe: {
    name?: string;
    value?: number[];
    display_value?: string;
    [k: string]: unknown;
  };
  /**
   * Impact time period. The value is UNIX time in seconds from epoch.
   */
  impact_timeframe: {
    name?: string;
    value?: number[];
    display_value?: string;
    [k: string]: unknown;
  };
  /**
   * Contributors
   */
  contributors: {
    name?: string;
    value?: string[];
    display_value?: string;
    [k: string]: unknown;
  };
  /**
   * Rights
   */
  rights?: {
    name?: string;
    value?: string[];
    excludes?: string[];
    display_value?: string;
    [k: string]: unknown;
  };
  [k: string]: unknown;
}